# OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: Project Management Orchestrator - Revised
  version: 2.0.0
  description: |
    Unified API with proper organization isolation and role-based access control.
    
    **Authentication**: All endpoints require Bearer token except /register and /login
    **Authorization**: Enforced at endpoint level based on user role
    **Data Isolation**: Users can only access data from their organization
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://api.example.com
    description: Production server

# ============================================
# AUTHENTICATION & ORGANIZATION MANAGEMENT
# ============================================
paths:
...
  /api/v1/performance/users/{user_id}/reviews:
    post:
      tags: [Performance]
      summary: Create performance review
      description: ⚠️ Only admin or manager can create reviews
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating, review_period_start, review_period_end]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comments:
                  type: string
                review_period_start:
                  type: string
                  format: date
                review_period_end:
                  type: string
                  format: date
      responses:
        '201':
          description: Review created
        '403':
          description: Only admin or manager can review
        '400':
          description: User not in organization or cannot review self

    get:
      tags: [Performance]
      summary: Get performance reviews
      description: Users can see own reviews, managers can see team reviews
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reviews
        '403':
          description: Cannot view other user's reviews
  /api/v1/performance/users/{user_id}/goals:
    post:
      tags: [Performance]
      summary: Create performance goal
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, target_date]
              properties:
                title:
                  type: string
                description:
                  type: string
                target_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Goal created

    get:
      tags: [Performance]
      summary: Get performance goals
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of goals
  /api/v1/performance/users/{user_id}/feedback:
    post:
      tags: [Performance]
      summary: Give peer feedback
      description: Any org member can give feedback to another
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feedback]
              properties:
                feedback:
                  type: string
                is_anonymous:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Feedback submitted
        '400':
          description: Cannot give feedback to self or user not in organization

    get:
      tags: [Performance]
      summary: Get peer feedback
      description: Users can see own feedback, managers can see team feedback
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of feedback

# ============================================
# DASHBOARD & ANALYTICS
# ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...